Assignment 2
You are given a Double Link List with one pointer of each node pointing to the next node just like in a single link list. The second pointer however points to any node in the list and not just the previous node. Write a program to duplicate [create a copy] this list.